+++ frontend/src/Page/shop/ShopDetail.jsx
@@ -73,6 +73,8 @@ function ShopDetail() {
         formData.append("cartCount", itemCount);
         formData.append("itemId", id);
         await postTokenData("/mypage/cart/add", formData);
+        alert('상품이 장바구니에 추가되었습니다.');
+        navigate('/shop');
     };
 
     const increaseCount = () => {
@@ -84,11 +86,14 @@ function ShopDetail() {
     };
 
     const onDelete = async () => {
+        const userConfirmed = window.confirm("정말로 삭제하시겠습니까?");
+        if (!userConfirmed) return;
         try {
             await deleteTokenData(`/shop/delete/${id}`);
+            alert('상품이 성공적으로 삭제되었습니다.');
             navigate("/shop", { replace: true });
         } catch (err) {
-            console.log(err);
+
         }
     };
 
@@ -108,7 +113,7 @@ function ShopDetail() {
                 const updatedItemFromDb = fetchItemResponse.data.item;
                 setLikeCount(updatedItemFromDb.likes);
             } catch (fetchError) {
-                console.error("상품 정보를 가져오는 데 오류 발생:", fetchError);
+
             }
         }
     };
